move bridge actions to own submodule
