+++ crates/astria-sequencer/src/checked_actions/bridge/bridge_unlock.rs
@@ -1,16 +1,14 @@
 use astria_core::{
     primitive::v1::{
         asset::IbcPrefixed,
-        TransactionId,
         ADDRESS_LEN,
     },
-    protocol::transaction::v1::action::{
-        BridgeLock,
-        BridgeTransfer,
-        BridgeUnlock,
-    },
+    protocol::transaction::v1::action::BridgeUnlock,
 };
 use astria_eyre::eyre::{
+    bail,
+    ensure,
+    OptionExt as _,
     Result,
     WrapErr as _,
 };
@@ -23,74 +21,83 @@ use tracing::{
     Level,
 };
 
-use super::{
-    bridge_lock::CheckedBridgeLockImpl,
-    bridge_unlock::CheckedBridgeUnlockImpl,
-    AssetTransfer,
+use crate::{
+    accounts::StateWriteExt as _,
+    address::StateReadExt as _,
+    bridge::{
+        StateReadExt as _,
+        StateWriteExt as _,
+    },
+    checked_actions::{
+        AssetTransfer,
+        TransactionSignerAddressBytes,
+    },
 };
-use crate::accounts::StateWriteExt as _;
 
+pub(crate) type CheckedBridgeUnlock = CheckedBridgeUnlockImpl<true>;
+
+/// This struct provides the implementation details of the checked bridge unlock action.
+///
+/// It is also used to perform checks on a bridge transfer action, which is essentially a cross
+/// between a bridge unlock and a bridge lock.
+///
+/// A `BridgeUnlock` action does not allow unlocking to a bridge account, whereas `BridgeTransfer`
+/// requires the `to` account to be a bridge one. Hence a bridge unlock is implemented via
+/// `CheckedBridgeUnlockImpl<true>` and has methods to allow checking AND executing the action,
+/// while the checks relevant to a bridge transfer are implemented via
+/// `CheckedBridgeUnlockImpl<false>`, where this has no method supporting execution.
 #[derive(Debug)]
-pub(crate) struct CheckedBridgeTransfer {
-    action: BridgeTransfer,
-    checked_bridge_unlock: CheckedBridgeUnlockImpl<false>,
-    checked_bridge_lock: CheckedBridgeLockImpl<false>,
+pub(crate) struct CheckedBridgeUnlockImpl<const PURE_UNLOCK: bool> {
+    action: BridgeUnlock,
+    tx_signer: TransactionSignerAddressBytes,
+    bridge_account_ibc_asset: IbcPrefixed,
 }
 
-impl CheckedBridgeTransfer {
+impl<const PURE_UNLOCK: bool> CheckedBridgeUnlockImpl<PURE_UNLOCK> {
     #[instrument(skip_all, err(level = Level::DEBUG))]
-    pub(super) async fn new<S: StateRead>(
-        action: BridgeTransfer,
+    pub(in crate::checked_actions) async fn new<S: StateRead>(
+        action: BridgeUnlock,
         tx_signer: [u8; ADDRESS_LEN],
-        tx_id: TransactionId,
-        position_in_tx: u64,
         state: S,
     ) -> Result<Self> {
-        let BridgeTransfer {
-            to,
-            amount,
-            fee_asset,
-            destination_chain_address,
-            bridge_address,
-            rollup_block_number,
-            rollup_withdrawal_event_id,
-        } = action.clone();
+        // TODO(https://github.com/astriaorg/astria/issues/1430): move stateless checks to the
+        // `BridgeUnlock` parsing.
+        ensure!(action.amount > 0, "amount must be greater than zero");
+        ensure!(
+            action.memo.len() <= 64,
+            "memo must not be more than 64 bytes"
+        );
+        ensure!(
+            !action.rollup_withdrawal_event_id.is_empty(),
+            "rollup withdrawal event id must be non-empty",
+        );
+        ensure!(
+            action.rollup_withdrawal_event_id.len() <= 256,
+            "rollup withdrawal event id must not be more than 256 bytes",
+        );
+        ensure!(
+            action.rollup_block_number > 0,
+            "rollup block number must be greater than zero",
+        );
 
-        let bridge_unlock = BridgeUnlock {
-            to,
-            amount,
-            memo: String::new(),
-            rollup_withdrawal_event_id,
-            rollup_block_number,
-            fee_asset: fee_asset.clone(),
-            bridge_address,
-        };
-        let checked_bridge_unlock =
-            CheckedBridgeUnlockImpl::<false>::new(bridge_unlock, tx_signer, &state)
-                .await
-                .wrap_err("failed to construct checked bridge unlock for bridge transfer")?;
+        state
+            .ensure_base_prefix(&action.to)
+            .await
+            .wrap_err("destination address has an unsupported prefix")?;
+        state
+            .ensure_base_prefix(&action.bridge_address)
+            .await
+            .wrap_err("source address has an unsupported prefix")?;
 
-        let bridge_lock = BridgeLock {
-            to,
-            amount,
-            asset: checked_bridge_unlock.bridge_account_ibc_asset().into(),
-            fee_asset,
-            destination_chain_address,
-        };
-        let checked_bridge_lock = CheckedBridgeLockImpl::<false>::new(
-            bridge_lock,
-            tx_signer,
-            tx_id,
-            position_in_tx,
-            &state,
-        )
-        .await
-        .wrap_err("failed to construct checked bridge lock for bridge transfer")?;
+        let bridge_account_ibc_asset = state
+            .get_bridge_account_ibc_asset(&action.bridge_address)
+            .await
+            .wrap_err("failed to get bridge account asset ID; account is not a bridge account")?;
 
         let checked_action = Self {
             action,
-            checked_bridge_unlock,
-            checked_bridge_lock,
+            tx_signer: tx_signer.into(),
+            bridge_account_ibc_asset,
         };
         checked_action.run_mutable_checks(state).await?;
 
@@ -98,80 +105,122 @@ impl CheckedBridgeTransfer {
     }
 
     #[instrument(skip_all, err(level = Level::DEBUG))]
-    pub(super) async fn run_mutable_checks<S: StateRead>(&self, state: S) -> Result<()> {
-        self.checked_bridge_unlock
-            .run_mutable_checks(&state)
+    pub(in crate::checked_actions) async fn run_mutable_checks<S: StateRead>(
+        &self,
+        state: S,
+    ) -> Result<()> {
+        if PURE_UNLOCK
+            && state
+                .is_a_bridge_account(&self.action.to)
+                .await
+                .wrap_err("failed to check if `to` address is a bridge account")?
+        {
+            bail!("bridge accounts cannot receive bridge unlocks");
+        }
+
+        let withdrawer = state
+            .get_bridge_account_withdrawer_address(&self.action.bridge_address)
             .await
-            .wrap_err("mutable checks for bridge unlock failed for bridge transfer")?;
-        self.checked_bridge_lock
-            .run_mutable_checks(&state)
+            .wrap_err("failed to get bridge account withdrawer address")?
+            .ok_or_eyre("bridge account must have a withdrawer address set")?;
+        ensure!(
+            *self.tx_signer.as_bytes() == withdrawer,
+            "signer is not the authorized withdrawer for the bridge account",
+        );
+
+        if let Some(existing_block_num) = state
+            .get_withdrawal_event_block_for_bridge_account(
+                &self.action.bridge_address,
+                &self.action.rollup_withdrawal_event_id,
+            )
             .await
-            .wrap_err("mutable checks for bridge lock failed for bridge transfer")
+            .wrap_err("failed to read withdrawal event block number from storage")?
+        {
+            bail!(
+                "withdrawal event ID `{}` used by block number {existing_block_num}",
+                self.action.rollup_withdrawal_event_id
+            );
+        }
+
+        Ok(())
     }
 
+    pub(in crate::checked_actions) fn action(&self) -> &BridgeUnlock {
+        &self.action
+    }
+
+    pub(super) fn record_withdrawal_event<S: StateWrite>(&self, mut state: S) -> Result<()> {
+        state
+            .put_withdrawal_event_block_for_bridge_account(
+                &self.action.bridge_address,
+                &self.action.rollup_withdrawal_event_id,
+                self.action.rollup_block_number,
+            )
+            .wrap_err("failed to write withdrawal event block number to storage")
+    }
+}
+
+impl CheckedBridgeUnlockImpl<true> {
     #[instrument(skip_all, err(level = Level::DEBUG))]
-    pub(super) async fn execute<S: StateWrite>(&self, mut state: S) -> Result<()> {
+    pub(in crate::checked_actions) async fn execute<S: StateWrite>(
+        &self,
+        mut state: S,
+    ) -> Result<()> {
         self.run_mutable_checks(&state).await?;
 
-        let from = &self.checked_bridge_unlock.action().bridge_address;
-        let to = &self.checked_bridge_unlock.action().to;
-        let asset = self.checked_bridge_unlock.bridge_account_ibc_asset();
-        let amount = self.checked_bridge_unlock.action().amount;
-
         state
-            .decrease_balance(from, asset, amount)
+            .decrease_balance(
+                &self.action.bridge_address,
+                &self.bridge_account_ibc_asset,
+                self.action.amount,
+            )
             .await
             .wrap_err("failed to decrease bridge account balance")?;
         state
-            .increase_balance(to, asset, amount)
+            .increase_balance(
+                &self.action.to,
+                &self.bridge_account_ibc_asset,
+                self.action.amount,
+            )
             .await
             .wrap_err("failed to increase destination account balance")?;
 
-        self.checked_bridge_lock.record_deposit(&mut state);
-        self.checked_bridge_unlock.record_withdrawal_event(state)
-    }
-
-    pub(super) fn action(&self) -> &BridgeTransfer {
-        &self.action
+        self.record_withdrawal_event(state)
     }
 }
 
-impl AssetTransfer for CheckedBridgeTransfer {
+impl CheckedBridgeUnlockImpl<false> {
+    pub(super) fn bridge_account_ibc_asset(&self) -> &IbcPrefixed {
+        &self.bridge_account_ibc_asset
+    }
+}
+
+impl AssetTransfer for CheckedBridgeUnlock {
     fn transfer_asset_and_amount(&self) -> Option<(IbcPrefixed, u128)> {
-        self.checked_bridge_lock.transfer_asset_and_amount()
+        Some((self.bridge_account_ibc_asset, self.action.amount))
     }
 }
 
+// NOTE: unit tests here cover only `CheckedBridgeUnlockImpl<true>`.  Test coverage of
+// `CheckedBridgeUnlockImpl<false>` is in `checked_actions::bridge_transfer`.
 #[cfg(test)]
-pub(super) mod tests {
+mod tests {
     use astria_core::{
-        primitive::v1::{
-            asset::{
-                Denom,
-                IbcPrefixed,
-            },
-            RollupId,
-        },
+        primitive::v1::RollupId,
         protocol::transaction::v1::action::*,
     };
 
-    use super::{
-        super::test_utils::address_with_prefix,
-        *,
-    };
+    use super::*;
     use crate::{
-        bridge::{
-            StateReadExt as _,
-            StateWriteExt as _,
-        },
         checked_actions::{
+            test_utils::address_with_prefix,
             CheckedBridgeSudoChange,
-            CheckedBridgeUnlock,
+            CheckedInitBridgeAccount,
         },
         test_utils::{
             assert_error_contains,
             astria_address,
-            dummy_bridge_transfer,
+            dummy_bridge_unlock,
             nria,
             Fixture,
             ASTRIA_PREFIX,
@@ -183,9 +232,9 @@ pub(super) mod tests {
     async fn should_fail_construction_if_amount_is_zero() {
         let fixture = Fixture::default_initialized().await;
 
-        let action = BridgeTransfer {
+        let action = BridgeUnlock {
             amount: 0,
-            ..dummy_bridge_transfer()
+            ..dummy_bridge_unlock()
         };
         let err = fixture
             .new_checked_action(action, *SUDO_ADDRESS_BYTES)
@@ -196,12 +245,28 @@ pub(super) mod tests {
     }
 
     #[tokio::test]
+    async fn should_fail_construction_if_memo_too_long() {
+        let fixture = Fixture::default_initialized().await;
+
+        let action = BridgeUnlock {
+            memo: ['a'; 65].into_iter().collect(),
+            ..dummy_bridge_unlock()
+        };
+        let err = fixture
+            .new_checked_action(action, *SUDO_ADDRESS_BYTES)
+            .await
+            .unwrap_err();
+
+        assert_error_contains(&err, "memo must not be more than 64 bytes");
+    }
+
+    #[tokio::test]
     async fn should_fail_construction_if_rollup_withdrawal_event_id_empty() {
         let fixture = Fixture::default_initialized().await;
 
-        let action = BridgeTransfer {
+        let action = BridgeUnlock {
             rollup_withdrawal_event_id: String::new(),
-            ..dummy_bridge_transfer()
+            ..dummy_bridge_unlock()
         };
         let err = fixture
             .new_checked_action(action, *SUDO_ADDRESS_BYTES)
@@ -215,9 +280,9 @@ pub(super) mod tests {
     async fn should_fail_construction_if_rollup_withdrawal_event_id_too_long() {
         let fixture = Fixture::default_initialized().await;
 
-        let action = BridgeTransfer {
+        let action = BridgeUnlock {
             rollup_withdrawal_event_id: ['a'; 257].into_iter().collect(),
-            ..dummy_bridge_transfer()
+            ..dummy_bridge_unlock()
         };
         let err = fixture
             .new_checked_action(action, *SUDO_ADDRESS_BYTES)
@@ -234,9 +299,9 @@ pub(super) mod tests {
     async fn should_fail_construction_if_rollup_block_number_is_zero() {
         let fixture = Fixture::default_initialized().await;
 
-        let action = BridgeTransfer {
+        let action = BridgeUnlock {
             rollup_block_number: 0,
-            ..dummy_bridge_transfer()
+            ..dummy_bridge_unlock()
         };
         let err = fixture
             .new_checked_action(action, *SUDO_ADDRESS_BYTES)
@@ -251,9 +316,9 @@ pub(super) mod tests {
         let fixture = Fixture::default_initialized().await;
 
         let prefix = "different_prefix";
-        let action = BridgeTransfer {
+        let action = BridgeUnlock {
             to: address_with_prefix([2; ADDRESS_LEN], prefix),
-            ..dummy_bridge_transfer()
+            ..dummy_bridge_unlock()
         };
         let err = fixture
             .new_checked_action(action, *SUDO_ADDRESS_BYTES)
@@ -271,9 +336,9 @@ pub(super) mod tests {
         let fixture = Fixture::default_initialized().await;
 
         let prefix = "different_prefix";
-        let action = BridgeTransfer {
+        let action = BridgeUnlock {
             bridge_address: address_with_prefix([50; ADDRESS_LEN], prefix),
-            ..dummy_bridge_transfer()
+            ..dummy_bridge_unlock()
         };
         let err = fixture
             .new_checked_action(action, *SUDO_ADDRESS_BYTES)
@@ -290,7 +355,7 @@ pub(super) mod tests {
     async fn should_fail_construction_if_bridge_account_not_initialized() {
         let fixture = Fixture::default_initialized().await;
 
-        let action = dummy_bridge_transfer();
+        let action = dummy_bridge_unlock();
         let err = fixture
             .new_checked_action(action, *SUDO_ADDRESS_BYTES)
             .await
@@ -303,10 +368,28 @@ pub(super) mod tests {
     }
 
     #[tokio::test]
+    async fn should_fail_construction_if_to_address_is_bridge_account() {
+        let mut fixture = Fixture::default_initialized().await;
+
+        let action = dummy_bridge_unlock();
+        fixture
+            .bridge_initializer(action.bridge_address)
+            .init()
+            .await;
+        fixture.bridge_initializer(action.to).init().await;
+        let err = fixture
+            .new_checked_action(action, *SUDO_ADDRESS_BYTES)
+            .await
+            .unwrap_err();
+
+        assert_error_contains(&err, "bridge accounts cannot receive bridge unlocks");
+    }
+
+    #[tokio::test]
     async fn should_fail_construction_if_withdrawer_address_not_set() {
         let mut fixture = Fixture::default_initialized().await;
 
-        let action = dummy_bridge_transfer();
+        let action = dummy_bridge_unlock();
         fixture
             .bridge_initializer(action.bridge_address)
             .with_no_withdrawer_address()
@@ -324,7 +407,7 @@ pub(super) mod tests {
     async fn should_fail_construction_if_signer_is_not_authorized() {
         let mut fixture = Fixture::default_initialized().await;
 
-        let action = dummy_bridge_transfer();
+        let action = dummy_bridge_unlock();
         fixture
             .bridge_initializer(action.bridge_address)
             .with_withdrawer_address([2; 20])
@@ -345,7 +428,7 @@ pub(super) mod tests {
     async fn should_fail_construction_if_withdrawal_event_id_already_used() {
         let mut fixture = Fixture::default_initialized().await;
 
-        let action = dummy_bridge_transfer();
+        let action = dummy_bridge_unlock();
         fixture
             .bridge_initializer(action.bridge_address)
             .init()
@@ -372,94 +455,60 @@ pub(super) mod tests {
     }
 
     #[tokio::test]
-    async fn should_fail_construction_if_destination_asset_not_same_as_source() {
+    async fn should_fail_execution_if_to_address_is_bridge_account() {
         let mut fixture = Fixture::default_initialized().await;
 
-        let action = dummy_bridge_transfer();
+        // Construct a checked bridge unlock while the `to` account is not a bridge account.
+        let action = dummy_bridge_unlock();
         fixture
             .bridge_initializer(action.bridge_address)
             .init()
             .await;
-        fixture
-            .bridge_initializer(action.to)
-            .with_asset(IbcPrefixed::new([10; 32]))
-            .init()
-            .await;
+        let to_address = action.to.bytes();
+        let checked_action: CheckedBridgeUnlock = fixture
+            .new_checked_action(action.clone(), *SUDO_ADDRESS_BYTES)
+            .await
+            .unwrap()
+            .into();
 
-        let err = fixture
-            .new_checked_action(action, *SUDO_ADDRESS_BYTES)
+        // Initialize the `to` account as a bridge account.
+        let init_bridge_account = InitBridgeAccount {
+            rollup_id: RollupId::new([2; 32]),
+            asset: "test".parse().unwrap(),
+            fee_asset: "test".parse().unwrap(),
+            sudo_address: None,
+            withdrawer_address: None,
+        };
+        let checked_init_bridge_account: CheckedInitBridgeAccount = fixture
+            .new_checked_action(init_bridge_account, to_address)
+            .await
+            .unwrap()
+            .into();
+        checked_init_bridge_account
+            .execute(fixture.state_mut())
+            .await
+            .unwrap();
+
+        // Try to execute the checked bridge unlock now - should fail due to `to` account now
+        // existing.
+        let err = checked_action
+            .execute(fixture.state_mut())
             .await
             .unwrap_err();
-
-        assert_error_contains(
-            &err,
-            "asset ID is not authorized for transfer to bridge account",
-        );
-    }
-
-    #[tokio::test]
-    async fn should_fail_construction_if_destination_rollup_id_not_found() {
-        let mut fixture = Fixture::default_initialized().await;
-
-        let action = dummy_bridge_transfer();
-        fixture
-            .bridge_initializer(action.bridge_address)
-            .init()
-            .await;
-        fixture
-            .bridge_initializer(action.to)
-            .with_no_rollup_id()
-            .init()
-            .await;
-
-        let err = fixture
-            .new_checked_action(action, *SUDO_ADDRESS_BYTES)
-            .await
-            .unwrap_err();
-
-        assert_error_contains(&err, "bridge lock must be sent to a bridge account");
-    }
-
-    #[tokio::test]
-    async fn should_fail_construction_if_asset_mapping_fails() {
-        let mut fixture = Fixture::default_initialized().await;
-
-        let action = dummy_bridge_transfer();
-        let asset = IbcPrefixed::new([10; 32]);
-        fixture
-            .bridge_initializer(action.bridge_address)
-            .with_asset(asset)
-            .init()
-            .await;
-        fixture
-            .bridge_initializer(action.to)
-            .with_asset(asset)
-            .init()
-            .await;
-
-        let err = fixture
-            .new_checked_action(action, *SUDO_ADDRESS_BYTES)
-            .await
-            .unwrap_err();
-
-        assert_error_contains(
-            &err,
-            "mapping from IBC prefixed bridge asset to trace prefixed not found",
-        );
+        assert_error_contains(&err, "bridge accounts cannot receive bridge unlocks");
     }
 
     #[tokio::test]
     async fn should_fail_execution_if_signer_is_not_authorized() {
         let mut fixture = Fixture::default_initialized().await;
 
-        // Construct a checked bridge transfer while the tx signer is the authorized withdrawer.
-        let action = dummy_bridge_transfer();
+        // Construct a checked bridge unlock while the tx signer is the authorized withdrawer.
+        let action = dummy_bridge_unlock();
         fixture
             .bridge_initializer(action.bridge_address)
             .init()
             .await;
-        fixture.bridge_initializer(action.to).init().await;
-        let checked_action: CheckedBridgeTransfer = fixture
+        let checked_action: CheckedBridgeUnlock = fixture
             .new_checked_action(action.clone(), *SUDO_ADDRESS_BYTES)
             .await
             .unwrap()
@@ -482,7 +531,7 @@ pub(super) mod tests {
             .await
             .unwrap();
 
-        // Try to execute the checked bridge transfer now - should fail due to tx signer no longer
+        // Try to execute the checked bridge unlock now - should fail due to tx signer no longer
         // being the authorized withdrawer.
         let err = checked_action
             .execute(fixture.state_mut())
@@ -498,49 +547,40 @@ pub(super) mod tests {
     async fn should_fail_execution_if_withdrawal_event_id_already_used() {
         let mut fixture = Fixture::default_initialized().await;
 
-        // Construct a checked bridge transfer while the withdrawal event ID is unused.
-        let action = dummy_bridge_transfer();
+        // Construct two checked bridge unlocks while the withdrawal event ID has not been used.
+        let action_1 = dummy_bridge_unlock();
+        let event_id = action_1.rollup_withdrawal_event_id.clone();
         fixture
-            .bridge_initializer(action.bridge_address)
+            .bridge_initializer(action_1.bridge_address)
             .init()
             .await;
-        fixture.bridge_initializer(action.to).init().await;
-        let checked_action: CheckedBridgeTransfer = fixture
-            .new_checked_action(action.clone(), *SUDO_ADDRESS_BYTES)
+        let checked_action_1: CheckedBridgeUnlock = fixture
+            .new_checked_action(action_1.clone(), *SUDO_ADDRESS_BYTES)
             .await
             .unwrap()
             .into();
-
-        // Execute a bridge unlock with the same withdrawal event ID.
-        let bridge_unlock = BridgeUnlock {
-            to: astria_address(&[3; ADDRESS_LEN]),
-            amount: 1,
-            fee_asset: nria().into(),
-            bridge_address: action.bridge_address,
-            memo: "a".to_string(),
-            rollup_block_number: 8,
-            rollup_withdrawal_event_id: action.rollup_withdrawal_event_id.clone(),
+        let action_2 = BridgeUnlock {
+            rollup_block_number: action_1.rollup_block_number.checked_add(1).unwrap(),
+            ..action_1.clone()
         };
-        let checked_bridge_unlock: CheckedBridgeUnlock = fixture
-            .new_checked_action(bridge_unlock.clone(), *SUDO_ADDRESS_BYTES)
+        let checked_action_2: CheckedBridgeUnlock = fixture
+            .new_checked_action(action_2.clone(), *SUDO_ADDRESS_BYTES)
             .await
             .unwrap()
             .into();
 
-        // Provide the bridge account with sufficient balance to execute the bridge unlock.
+        // Execute the first bridge unlock to write the withdrawal event ID to state. Need to
+        // provide the bridge account with sufficient balance.
         fixture
             .state_mut()
-            .increase_balance(&bridge_unlock.bridge_address, &nria(), bridge_unlock.amount)
+            .increase_balance(&action_1.bridge_address, &nria(), action_1.amount)
             .await
             .unwrap();
-        checked_bridge_unlock
-            .execute(fixture.state_mut())
-            .await
-            .unwrap();
+        checked_action_1.execute(fixture.state_mut()).await.unwrap();
 
-        // Try to execute the checked bridge transfer now with the same withdrawal event ID - should
-        // fail due to the ID being used already.
-        let err = checked_action
+        // Try to execute the second checked bridge unlock now with the same withdrawal event ID -
+        // should fail due to the ID being used already.
+        let err = checked_action_2
             .execute(fixture.state_mut())
             .await
             .unwrap_err();
@@ -548,8 +588,8 @@ pub(super) mod tests {
         assert_error_contains(
             &err,
             &format!(
-                "withdrawal event ID `{}` used by block number {}",
-                action.rollup_withdrawal_event_id, bridge_unlock.rollup_block_number
+                "withdrawal event ID `{event_id}` used by block number {}",
+                action_1.rollup_block_number
             ),
         );
     }
@@ -558,13 +598,12 @@ pub(super) mod tests {
     async fn should_fail_execution_if_bridge_account_has_insufficient_balance() {
         let mut fixture = Fixture::default_initialized().await;
 
-        let action = dummy_bridge_transfer();
+        let action = dummy_bridge_unlock();
         fixture
             .bridge_initializer(action.bridge_address)
             .init()
             .await;
-        fixture.bridge_initializer(action.to).init().await;
-        let checked_action: CheckedBridgeTransfer = fixture
+        let checked_action: CheckedBridgeUnlock = fixture
             .new_checked_action(action.clone(), *SUDO_ADDRESS_BYTES)
             .await
             .unwrap()
@@ -581,22 +620,14 @@ pub(super) mod tests {
     async fn should_execute() {
         let mut fixture = Fixture::default_initialized().await;
 
-        // Construct the checked bridge transfer while the account has insufficient balance to
-        // ensure balance checks are only part of execution.
-        let action = dummy_bridge_transfer();
-        let rollup_id = RollupId::new([7; 32]);
-        let asset = nria();
+        // Construct the checked bridge unlock while the account has insufficient balance to ensure
+        // balance checks are only part of execution.
+        let action = dummy_bridge_unlock();
         fixture
             .bridge_initializer(action.bridge_address)
-            .with_asset(asset.clone())
             .init()
             .await;
-        fixture
-            .bridge_initializer(action.to)
-            .with_rollup_id(rollup_id)
-            .init()
-            .await;
-        let checked_action: CheckedBridgeTransfer = fixture
+        let checked_action: CheckedBridgeUnlock = fixture
             .new_checked_action(action.clone(), *SUDO_ADDRESS_BYTES)
             .await
             .unwrap()
@@ -632,30 +663,5 @@ pub(super) mod tests {
             .await
             .unwrap();
         assert_eq!(rollup_block_number, Some(action.rollup_block_number));
-
-        // Check the deposit is recorded.
-        let deposits = fixture
-            .state()
-            .get_cached_block_deposits()
-            .get(&rollup_id)
-            .unwrap()
-            .clone();
-        assert_eq!(deposits.len(), 1);
-        let deposit = &deposits[0];
-        assert_eq!(deposit.bridge_address, action.to);
-        assert_eq!(deposit.rollup_id, rollup_id);
-        assert_eq!(deposit.amount, action.amount);
-        assert_eq!(deposit.asset, Denom::from(asset));
-        assert_eq!(
-            deposit.destination_chain_address,
-            action.destination_chain_address
-        );
-        assert_eq!(deposit.source_transaction_id.get(), [11; 32]);
-        assert_eq!(deposit.source_action_index, 11);
-
-        // Check the deposit event is cached.
-        let deposit_events = fixture.into_events();
-        assert_eq!(deposit_events.len(), 1);
-        assert_eq!(deposit_events[0].kind, "tx.deposit");
     }
 }
